refactor(flight): split recording finalization + terminal-state into ParsekFlight.Finalization.cs
Verbatim partial-class move of MaybeShowPendingTreeMergeDialogOnFlightReady, ShouldEnsureActiveRecordingTerminalState, ResolveFinalizationCacheForRecording, EnsureActiveRecordingTerminalState, the FinalizationLiveVesselAccess nested class, and FinalizeIndividualRecording. Updated the Bug278 source-text pin to scan all ParsekFlight*.cs partial siblings since two pinned strings now live in the new partial.
